requirements.txt
flask
flask-cors
python-dotenv
better-profanity
pymongo
python-dotenv
pymongo[srv]